[Django] Course-store/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-crispy-forms==1.14.0
django-mysql==3.8.1
django-seed==0.3.1
djangorestframework==3.13.1
Faker==14.2.0
gunicorn==20.1.0
mysqlclient==2.0.1
Pillow==8.4.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
toposort==1.7
